+++ src/ios/CDVBackgroundGeolocation.m
@@ -193,7 +193,7 @@
         CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:params];
         [commandDelegate sendPluginResult:result callbackId:command.callbackId];
     } failure:^(NSString* error) {
-        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:error];
+        CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error];
         [commandDelegate sendPluginResult:result callbackId:command.callbackId];
     }];
 }
@@ -770,6 +770,24 @@
 {
     BOOL isPowerSaveMode = [bgGeo isPowerSaveMode];
     CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:isPowerSaveMode];
+    [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+}
+
+- (void) isIgnoringBatteryOptimizations:(CDVInvokedUrlCommand *) command
+{
+    CDVPluginResult *result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:NO];
+    [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+}
+
+- (void) requestSettings:(CDVInvokedUrlCommand *) command
+{
+    CDVPluginResult* result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"No iOS Implementation"];
+    [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+}
+
+- (void) showSettings:(CDVInvokedUrlCommand *) command
+{
+    CDVPluginResult* result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"No iOS Implementation"];
     [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
 }